requirements.txt
psycopg2-binary
xlwt
GitPython
SQLAlchemy
imbalanced-learn
imblearn
matplotlib
numpy
pandas
python-dateutil
scipy
seaborn
sklearn
